requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
dnspython==2.5.0
flasgger==0.9.7.1
Flask==3.0.2
Flask-RESTful==0.3.10
flask-swagger==0.2.14
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
six==1.16.0
Werkzeug==3.0.1
